requirements.txt
django>=5.0
django-environ
python-dotenv
groq
google-genai
ollama
urllib3
gunicorn
